+++ lock.h
@@ -1,0 +1,4 @@
+#ifndef LOCK_H
+#define LOCK_H
+void lock_start(void);
+#endif