+++ CHANGELOG.md
@@ -12,6 +12,10 @@ There are 6 types of changes :
 
 # Changes
 
+## 2023-05-20
+### Added
+*   I added a great lock screen
+
 ## 2023-05-18
 ### Changed
 *   I changed multi spaces indentation by tab